+++ src/main/java/com/norab/actor/ActorDao.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 
 @Component
 public interface ActorDao<Actor> {
-    List<Movie> allMoviesByActor(Integer id);
+    List<Movie> allMoviesByActor(Long id);
     List<Actor> selectActors();
     int insertActor(Actor actor);
-    int deleteActor(int id);
-    Optional<Actor> selectActorById(int id);
-    int updateActor(int id, Actor actor);
+    int deleteActor(Long id);
+    Optional<Actor> selectActorById(Long id);
+    int updateActor(Long id, Actor actor);
 }
