+++ src/main/java/com/norab/movie/MovieDao.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 public interface MovieDao<Movie> {
     List<Movie> selectMovies();
     int insertMovie(Movie movie);
-    int deleteMovie(int id);
-    Optional<Movie> selectMovieById(int id);
-    int updateMovie(int id, Movie movie);
+    int deleteMovie(Long id);
+    Optional<Movie> selectMovieById(Long id);
+    int updateMovie(Long id, Movie movie);
 }
