+++ src/main/java/com/norab/role/RoleDao.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 public interface RoleDao<Plays> {
     List<Plays> selectRoles();
     int insertRole(Plays plays);
-    int deleteRole(int id);
-    Optional<Plays> selectRoleById(int id);
-    int updateRole(int id, Plays plays);
+    int deleteRole(Long id);
+    Optional<Plays> selectRoleById(Long id);
+    int updateRole(Long id, Plays plays);
 }
