+++ src/main/java/com/norab/role_photo/RolePhotoDao.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 public interface RolePhotoDao<RolePhoto> {
     List<RolePhoto> selectRolePhotos();
     int insertRolePhoto(RolePhoto rolePhoto);
-    int deleteRolePhoto(Integer id);
-    Optional<RolePhoto> selectRolePhotoById(Integer id);
-    int updateRolePhoto(Integer id, RolePhoto rolePhoto);
+    int deleteRolePhoto(Long id);
+    Optional<RolePhoto> selectRolePhotoById(Long id);
+    int updateRolePhoto(Long id, RolePhoto rolePhoto);
 }
